refactor: 공통으로 뺼수 있는거 리펙토링 view, controller 및 메세지 다른 파일로 분리
